+++ wayflowcore/tests/run_tests.sh
@@ -14,8 +14,8 @@ REPO_ROOT=$( cd "$(dirname "${BASH_SOURCE[0]}")"/.. ; pwd -P )
 if [ "$1" = "--parallel" ]; then
     # Run tests in parallel
     # disable a2a server tests (issue #40)  # tests/a2a
-    pytest tests/test_docstring.py tests/datastores tests/agentserver tests/transforms/test_summarization_transforms.py
-    pytest -n auto $REPO_ROOT/tests --dist loadscope --ignore=tests/datastores/ --ignore=tests/test_docstring.py --ignore=tests/agentserver/ --ignore=tests/transforms/test_summarization_transforms.py --ignore=tests/a2a
+    pytest tests/test_docstring.py tests/datastores tests/agentserver tests/transforms/test_summarization_transforms.py tests/mcptools/test_mcp_tools.py
+    pytest -n auto $REPO_ROOT/tests --dist loadscope --ignore=tests/datastores/ --ignore=tests/test_docstring.py --ignore=tests/agentserver/ --ignore=tests/transforms/test_summarization_transforms.py --ignore=tests/a2a --ignore=tests/mcptools/test_mcp_tools.py
 else
     # Run tests normally
     pytest $REPO_ROOT/tests